test(runtime): isolate SYS(16) stack fixtures
