+++ tests/CMakeLists.txt
@@ -5641,6 +5641,7 @@ copperfin_add_test(test_prg_engine_runtime_surface_functions cf_xbase_runtime
     test_prg_engine_runtime_surface_functions_events_02.cpp
     test_prg_engine_runtime_surface_functions_events_03.cpp
     test_prg_engine_runtime_surface_functions_expressions_01.cpp
+    test_prg_engine_runtime_surface_functions_sys16.cpp
     test_prg_engine_runtime_surface_functions_sys2012.cpp
     test_prg_engine_runtime_surface_functions_sys2030.cpp
     test_prg_engine_runtime_surface_functions_sys2021.cpp
